development/app/pyproject.toml
[tool.poetry]
name = "wagtail-ninja-dev"
version = "1.0.0"
description = "wagtail-ninja-dev"
authors = ["Dan Brosnan <dpjbrosnan@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.12"
Django = "~4.2"
django-redis = "~5.4"
wagtail-ninja = { path = "/wagtail_ninja", develop = true }
text-unidecode = "^1.3"
psycopg = "^3.2.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest-django = "~4.5"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
coverage = "~7.2"
pre-commit = "~3.3"
black = "^24.4.2"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
